[SC-171] Make SC service actions sticky (#129)
This is a workaround for the  problem that arises from product versions
being replaced. To date we have found no other way to force updates than
to replace product versions. However, we are focused on doing so because
we’re developing. In production, we should release a version and freeze it,
and there is no problem with updates when adding a new version (appending
a new item to the version list).

This PR depends on PR https://github.com/Sage-Bionetworks/service-catalog-library/pull/192