[SC-192] Refactor references to product templates (#142)
New tags were created in https://github.com/Sage-Bionetworks/service-catalog-library/tags
so now we can properly reference from a tagged directory of files in S3

The versions v1.0.0, v1.0.1, and v1.0.2 are all identical.  The idea
behind creating idential versions is so we can match the version of the
S3 bucket folder with the version on the file.  If we didn't create
multiple versions we would have to reference with `v1.0.0` for the
folder and `sc-ec2-linux-jumpcloud-v1.0.2.yaml` for the file which
would be pretty confusing. i.e.:
```
- Description: !Sub 'Fix to support multiple portfolios. Last updated at ${StackDatetime}.'
  Info:
    LoadTemplateFromURL: !Sub '${RepoRootURL}/v1.0.0/ec2/sc-ec2-linux-jumpcloud-v1.0.2.yaml'
  Name: 'v1.0.2'
```

Also I didn't wanted to keep the existing references the same as much as
possible otherwise AWS may create new product versions.

On the next version we will no longer have version info in the file.  it
will be something like this..
```
- Description: !Sub 'Fix to support multiple portfolios. Last updated at ${StackDatetime}.'
  Info:
    LoadTemplateFromURL: !Sub '${RepoRootURL}/v1.1.1/ec2/sc-ec2-linux-jumpcloud.yaml'
  Name: 'v1.1.1'
```