2nd attempt at DRY tag options
This is a 2nd attempt at commit a7f72bd6b3. The first attempt failed
because using the Sceptre file_contents resolver[1] passed the file
content as a string to the parameter. The jinja template that processes
the parameter requires a list object.  The new and improved
Sceptre file resolver[2] fixes this by getting the contents of a json or
yaml file and passing it to the parameter as an object.

[1] https://sceptre.cloudreach.com/2.2.1/docs/resolvers.html#file-contents
[2] https://github.com/Sage-Bionetworks-IT/sceptre-file-resolver

depends on https://github.com/Sage-Bionetworks-IT/sceptre-file-resolver/pull/1